requirements.txt
streamlit
google-generativeai
python-dotenv
streamlit-mic-recorder